+++ cloud/google/logs.go
@@ -1,6 +1,7 @@
 package google
 
 import (
+	"errors"
 	"io"
 
 	pb "github.com/datacol-io/datacol/api/models"
@@ -9,4 +10,8 @@ import (
 
 func (g *GCPCloud) LogStream(app string, w io.Writer, opts pb.LogStreamOptions) error {
 	return sched.LogStreamReq(g.kubeClient(), w, g.DeploymentName, app, opts)
+}
+
+func (g *GCPCloud) DockerCredsGet() (*pb.DockerCred, error) {
+	return nil, errors.New("not implemented")
 }