+++ hack/make.rb
@@ -1,6 +1,6 @@
 require "rubygems"
 
-$version = ENV.fetch("VERSION", "1.0.0-alpha.15")
+$version = ENV.fetch("VERSION", "1.0.0-alpha.16")
 $env = ENV.fetch("DATACOL_ENV") # dev or prod
 $cgo = ENV.fetch("CGO_ENABLED", "0")
 $rb_token = ENV.fetch("ROLLBAR_TOKEN")
@@ -99,6 +99,8 @@ def with_cmd(cmd)
     exit(code)
   end
 end
+
+with_cmd("gcloud config set account dinesh1042@gmail.com")
 
 if ARGV.size > 0
   send(ARGV[0])